movies_admin/requirements/base.txt
asgiref==3.4.1
Django==3.2.6
django-model-utils==4.0.0
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.8.0
gunicorn==20.1.0
loguru==0.5.3
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tqdm==4.62.3
